requirements.txt
wheel>=0.38.0
setuptools>=65.5.1
requests
xmltodict
vectra_api_tools >=2.4
pyvmomi
pyVim
PyJWT
cryptography
urllib3
libtaxii
python-dateutil
pip
colorlog
keyring
pytz
cabby
stix
six
furl
orderedmultidict
lxml
mixbox
cybox
certifi
idna
pycparser
zipp
weakrefmethod
cachetools
python-dotenv
mcafee-epo
